Tables: move DepartmentGateway class to Departments namespace
